edges values changed to int

+++ src/util/Utils.h
@@ -10,9 +10,9 @@
 class Utils {
 
 public:
-    static std::vector<std::vector<float>> loadGEucSpace(const std::string &file_path);
+    static std::vector<std::vector<int>> loadGEucSpace(const std::string &file_path);
 
-    static std::vector<std::vector<float>> readVList(const std::string &file_path);
+    static std::vector<std::vector<int>> readVList(const std::string &file_path);
 
     static float stdDev(std::vector<float> &items, float average);
